@@ -8,7 +8,7 @@ authors = [{name = "Bastien GANDOUET", email = "bastien@mozaiqu.es"}]
 license = {file = "LICENSE"}
 classifiers = ["License :: OSI Approved :: MIT License"]
 dynamic = ["version"]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
     "smart_open[s3] >=6.0.0,<7.0.0",
     "toml == 0.10.2",
